Rebuild Serializer based on JSON:API Serializer
This is the start of switching to a serialization class that will
handle translating SensorThings API responses into valid JSON:API
for use by Ember Data.

This first commit will parse collection responses from SensorThings
API. It maps the entities into data objects, with attributes, links,
and relationships.

Still to do:

* Handle getting a single entity from STA as a response
* Use the `@iot.count` attribute for meta collection information
* Move entities included using `$expand` into the `included` array
  used in JSON:API
* Refactor the Serializer class to separate concerns for methods,
  which will be used as the Serializer is expanded to handle more
  cases

(The reverse case of serializing JSON:API data for sending data to
SensorThings API is not planned at this time, but should be doable
with the Serializer class. JSON:API seems close enough to OData to
work.)