Convert relationship keys for Ember Data
When specifying the relationship entities in JSON:API, the keys
must match the same keys used in the Ember Data models by using
`camelCase` instead of `PascalCase` (as used in SensorThings API).
By making this change, Ember Data now can automatically know what
URL to query to fetch related resources for a model; in this case,
asking for a Thing's "locations" will automatically issue an AJAX
request to SensorThings API and bring the Location entities into
the Ember Data store.