Parse embedded entities from SensorThings API
If a request to the Ember Data store uses $expand, then that param
will be sent to SensorThings API. This change updates the serializer
to parse the embedded entity into valid JSON:API format, causing the
embedded entities to be automatically ingested into the Ember Data
store.

This plus the previous commit allows all the Things and Locations
to be loaded onto the map as markers with only 2 XHR requests.