Pass all custom URL query parameters
Instead of limiting the query parameters you can send, all custom
options specified in a query to the Ember Data Store are passed to
SensorThings API. This is a more appropriate way of passing the
options.

The `getNextPage` method has been updated to support passing the
limit and total options recursively instead of modifying the query
options object, which prevents them from being added to the URL
inadvertently.

The `setUrlQueryParams` has been removed as it has been replaced.