Fix serialization of single vs multiple
Old method was to serialize single entity and entity collections
from SensorThings API into an *array* for JSON:API. This is not
correct, as Ember Data expects a *single* object instead of an
array for model relationships that are one-to-one (and similar).

This commit fixes that by serializing single entities to an object,
and serializing collections of entities to arrays.