Fix loading of air temperature datastream
New method is to load all the datastreams (with embedded Observed
Properties) and filter on the client side. This commit does that
for the "Air Temperature" property.

It seems to be very important to use a Promise Object, as otherwise
an empty model is sent to the component and no model properties
ever render. I need to research more to find out if this is the best
way to do this in Ember.

The live panel component has been changed to display the "Loading"
version while the Datastream promise is waiting to resolve.