Fix "Last Update" display
The display of the phenomenon time for the latest Observation in the
live panel now uses the updated interface for accessing the live
panel Datastreams. Promises are extensively used to handle async
loading of Observations. Note that PromiseObject must be used when
a component needs to render a property from an asynchronously loaded
model.