Fix chart not shrinking on window resize
When the chart is visible by selecting a station, making the browser
window larger would cause the chart to resize to the new width. But
making the window smaller would not shrink the chart.

This commit fixes that by adding an event handler to check for the
window being resized and to automatically grow/shrink the chart to
its parent container size.

Additionally it fixes the "Loading..." element in the chart from
being visible as a white rectangle behind the chart when resizing
the window.