Switch map tiles to ArcticWebMap v2.0
Adds support for custom projection XYZ map tiles that follow the
OpenStreetMap tile hosting scheme (that is, the tile order axis).

Not sure if I will add support for WMS/WMTS as it may not be needed.
I may test the loading speed and bandwidth usage of WMTS vs XYZ
eventually.