Add Overlay Popup when selecting a station marker
If only a single station is at the click, then the station name
will be displayed in the popup and the station data loaded into
the rest of the dashboard.

If multiple stations are near the click, such as in the case that
stations are very close or overlapping, then a warning is printed
to the console. I will add a custom popup that lets the user select
the station they want or zoom to the location.