Display multiple options in feature popup
If multiple features are close to the point selected, display a
popup that lets the user select an option from a list. This works
even when multiple features are at the same exact location.

If there are more than a dozen options, it may look bad in the popup
and another solution may be needed.