Add "Zoom Closer" option in Popups
This lets the user inspect the region around a station, or to see
a group of stations closer to pick the correct one. Zooming closer
will close the existing popup, as a popup with multiple options is
no longer correct when the features are farther apart (such as after
the zoom in).

The zoom animation is limited to zoom level 14, which is enough to
see the surrounding area without getting down to a level without any
context (such as < 10 metres). 50px padding is configured so that
features aren't hidden along the border of the map. EaseOut is used
to make it easier to see what the animation is doing.